Revert "feat: filter nearby toilets"
