+++ src/components/useNearbyLoos.js
@@ -9,14 +9,6 @@ const NEARBY_LOOS_QUERY = gql`
         lat
         lng
       }
-      fee
-      type
-      access
-      attended
-      automatic
-      accessibleType
-      babyChange
-      radar
     }
   }
 `;
