+++ src/pages/HomePage.js
@@ -1,9 +1,10 @@
-import React, { useState } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import { useParams } from 'react-router-dom';
 import { loader } from 'graphql.macro';
 import { useQuery } from '@apollo/client';
 
+import config from '../config';
 import PageLayout from '../components/PageLayout';
 import LooMap from '../components/LooMap';
 import useMapPosition from '../components/useMapPosition';
@@ -11,26 +12,10 @@ import useNearbyLoos from '../components/useNearbyLoos';
 import Box from '../components/Box';
 import ToiletDetailsPanel from '../components/ToiletDetailsPanel';
 
-import config, { FILTERS_KEY } from '../config';
-
 const FIND_BY_ID = loader('./findLooById.graphql');
 
 const HomePage = ({ initialPosition, ...props }) => {
   const [mapPosition, setMapPosition] = useMapPosition(config.fallbackLocation);
-
-  let initialState = config.getSettings(FILTERS_KEY);
-
-  // default any unsaved filters as 'false'
-  config.filters.forEach((filter) => {
-    initialState[filter.id] = initialState[filter.id] || false;
-  });
-
-  const [filters, setFilters] = useState(initialState);
-
-  // keep local storage and state in sync
-  React.useEffect(() => {
-    window.localStorage.setItem(FILTERS_KEY, JSON.stringify(filters));
-  }, [filters]);
 
   React.useEffect(() => {
     // Set the map position if initialPosition prop exists
@@ -41,7 +26,7 @@ const HomePage = ({ initialPosition, ...props }) => {
     }
   }, [initialPosition, setMapPosition]);
 
-  const { data: toiletData } = useNearbyLoos({
+  const { data: loos } = useNearbyLoos({
     variables: {
       lat: mapPosition.center.lat,
       lng: mapPosition.center.lng,
@@ -58,45 +43,11 @@ const HomePage = ({ initialPosition, ...props }) => {
     },
   });
 
-  // get the filter objects from config for the filters applied by the user
-  const applied = config.filters.filter((filter) => filters[filter.id]);
-
-  // restrict the results to only those toilets which pass all of our filter requirements
-  const toilets = toiletData.filter((toilet) =>
-    applied.every((filter) => {
-      const field = filter.field || filter.id;
-      const value = toilet[field];
-
-      if (field === 'fee') {
-        return !value;
-      }
-
-      if (value === null) {
-        return false;
-      }
-
-      switch (field) {
-        case 'type':
-          return value === 'UNISEX';
-
-        case 'accessibleType':
-          return value !== 'NONE';
-
-        default:
-          return !!value;
-      }
-    })
-  );
-
   return (
-    <PageLayout
-      filters={filters}
-      onFilterChange={setFilters}
-      onSelectedItemChange={(center) => setMapPosition({ center })}
-    >
+    <PageLayout onSelectedItemChange={(center) => setMapPosition({ center })}>
       <Box height="100%" display="flex" position="relative">
         <LooMap
-          loos={toilets}
+          loos={loos}
           center={mapPosition.center}
           zoom={mapPosition.zoom}
           onViewportChanged={setMapPosition}
